Adicao dos metodos post e delete

+++ src/main/java/com/guerra/domain/Aluno.java
@@ -13,23 +13,23 @@ public class Aluno implements Serializable {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+	private Integer id;
 	private String nome;
 	
 	public Aluno() {
 	}
 
-	public Aluno(Long id, String nome) {
+	public Aluno(Integer id, String nome) {
 		super();
 		this.id = id;
 		this.nome = nome;
 	}
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
